requirements1.txt
absl-py==1.0.0
astor==0.8.1
Babel==2.9.1
bce-python-sdk==0.8.64
cachetools==4.2.4
certifi==2021.10.8
cfgv==3.3.1
charset-normalizer==2.0.10
click==8.0.3
colorama==0.4.4
colorlog==6.6.0
cssselect==1.1.0
cssutils==2.3.0
cycler==0.11.0
Cython==0.29.26
decorator==5.1.1
dill==0.3.4
distlib==0.3.4
easyocr==1.4.1
et-xmlfile==1.1.0
filelock==3.4.2
flake8==4.0.1
Flask==2.0.2
Flask-Babel==2.0.0
fonttools==4.28.5
future==0.18.2
google-auth==2.3.3
google-auth-oauthlib==0.4.6
grpcio==1.43.0
h5py==3.6.0
identify==2.4.4
idna==3.3
imageio==2.9.0
imgaug==0.4.0
importlib-metadata==4.10.1
itsdangerous==2.0.1
jieba==0.42.1
Jinja2==3.0.3
joblib==1.1.0
kiwisolver==1.3.2
lmdb==1.3.0
lxml==4.7.1
Markdown==3.3.6
MarkupSafe==2.0.1
matplotlib==3.5.1
mccabe==0.6.1
multiprocess==0.70.12.2
networkx==2.6.3
nodeenv==1.6.0
numpy==1.19.3
oauthlib==3.1.1
opencv-contrib-python
openpyxl==3.0.9
packaging==21.3
paddlenlp==2.2.3
paddleocr==2.4
paddlepaddle==2.2.2
pandas==1.3.5
Pillow==8.2.0
platformdirs==2.4.1
pre-commit==2.17.0
premailer==3.10.0
protobuf==3.19.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyclipper==1.3.0.post2
pycodestyle==2.8.0
pycryptodome==3.12.0
pyflakes==2.4.0
pyparsing==3.0.7
python-bidi==0.4.2
python-dateutil==2.8.2
python-Levenshtein==0.12.2
pytz==2021.3
PyWavelets==1.2.0
PyYAML==6.0
requests==2.27.1
requests-oauthlib==1.3.0
rsa==4.8
scikit-image==0.19.1
scikit-learn==1.0.2
scipy==1.7.3
seaborn==0.11.2
seqeval==1.2.2
Shapely==1.8.0
shellcheck-py==0.8.0.3
six==1.16.0
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
thop==0.0.31.post2005241907
threadpoolctl==3.0.0
tifffile==2021.11.2
toml==0.10.2
torch==1.10.1
torchvision==0.11.2
tqdm==4.62.3
typing_extensions==4.0.1
urllib3==1.26.8
virtualenv==20.13.0
visualdl==2.2.3
Werkzeug==2.0.2
wincertstore==0.2
zipp==3.7.0
